requirements.txt
asgiref==3.4.1
astroid==2.3.3
Django==3.2
django-cors-headers==3.8.0
django-environ==0.7.0
djangorestframework==3.12.4
docutils==0.15.2
gunicorn==20.1.0
hyperlink==20.0.1
idna==3.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mock==4.0.3
Pillow==8.3.2
psycopg2-binary==2.8.6
py==1.8.0
pylint==2.4.2
pytz==2021.1
six==1.15.0
sqlparse==0.4.2
whitenoise==5.3.0
wrapt==1.11.2
